Divide code into separate files
