+++ Runtime/LifetimeScopeExtension.cs
@@ -1,0 +1,16 @@
+using System;
+
+namespace UniMob
+{
+    public static class LifetimeScopeExtension
+    {
+        // ReSharper disable once ParameterOnlyUsedForPreconditionCheck.Global
+        public static void ThrowIfDisposed(ILifetimeScope scope)
+        {
+            if (scope.Lifetime.IsDisposed)
+            {
+                throw new ObjectDisposedException("Lifetime is disposed");
+            }
+        }
+    }
+}
