+++ Runtime/LifetimeScopeExtension.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: a841d77d90c34941a5bb1474397be87e
+timeCreated: 1693904156
